Quickly convert Markdown to LaTeX or vice versa from your clipboard with Pandoc.
A simple routine to convert snippets of markdown to LateX from your
clipboard or vice versa. The workflow creates either a *.md or *.tex
file relative to your input and at the location you choose. The file is
converted with Pandoc by calling it from the terminal. The converted
file subsequently gets output at the same place. Existing files will be
overwritten.

Pandoc is key and has to be installed:

	$ brew install pandoc

More information: https://pandoc.org/